backend/telegram-scrapper/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
telethon
